lucommon/requirements.txt
Django==1.8.1
MySQL-python==1.2.5
PyMySQL==0.7.2
PyYAML==3.11
argparse==1.2.1
defusedxml==0.4.1
django-cors-headers==1.1.0
django-filter==0.9.2
django-filters==0.2.1
django-logging-json==1.4
django-redis-cache==1.6.5
django-rest-swagger==0.3.5
django-simple-sso==0.9.3
djangorestframework==3.3.3
djangorestframework-xml==1.3.0
djangorestframework-yaml==1.0.3
drf-extensions==0.2.8
itsdangerous==0.24
lxml==3.6.0
redis==2.10.5
requests==2.10.0
six==1.10.0
termcolor==1.1.0
uWSGI==2.0.12
webservices==0.6
wsgiref==0.1.2
django-reversion==1.10.0
django-reversion-compare==0.6.2
